ajout position producer

+++ server/src/main/java/com/xoff/chessvger/chess/board/CoupleZobristMaterial.java
@@ -7,12 +7,12 @@ import lombok.Getter;
 public class CoupleZobristMaterial {
 
 
-  private final int zobrist;
+  private final long zobrist;
 
   private final long material;
 
 
-  public CoupleZobristMaterial(int zobrist, long material) {
+  public CoupleZobristMaterial(long zobrist, long material) {
     this.zobrist = zobrist;
     this.material = material;
   }
